requirements.txt
python-dotenv
requests
pandas
apache-airflow==2.10.1
psycopg2-binary